[Parser]fix CP-18976, floor infer dtype issue
(cherry picked from commit 3b12a595fc61c84b5350d0979aedc59cdbc78511)

+++ AIPUBuilder/Parser/ops/onnx_ops/math_ops.py
@@ -600,7 +600,7 @@ class FloorOp(LayoutUnawareOp, OpHasOneOutPort, OnnxOp):
     def infer_shape(self):
         super(FloorOp, self).infer_shape()
         inputs = self.get_input_tensors()
-        out_tensor = np.floor(inputs[0]).astype(np.float32)
+        out_tensor = np.floor(inputs[0])
         self.set_out_tensor(out_tensor)
